Update validate config file in rasa train.
