+++ rasa/cli/train.py
@@ -3,8 +3,21 @@ import tempfile
 from typing import List, Optional, Text, Dict
 import rasa.cli.arguments as arguments
 
-from rasa.cli.utils import get_validated_path
-from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DATA_PATH, DEFAULT_DOMAIN_PATH
+from rasa.cli.utils import (
+    get_validated_path,
+    missing_config_keys,
+    print_error,
+    print_warning,
+)
+from rasa.constants import (
+    DEFAULT_CONFIG_PATH,
+    DEFAULT_DATA_PATH,
+    DEFAULT_DOMAIN_PATH,
+    FALLBACK_CONFIG_PATH,
+    CONFIG_MANDATORY_KEYS_NLU,
+    CONFIG_MANDATORY_KEYS_CORE,
+    CONFIG_MANDATORY_KEYS,
+)
 
 
 # noinspection PyProtectedMember
@@ -52,7 +65,8 @@ def train(args: argparse.Namespace) -> Optional[Text]:
     domain = get_validated_path(
         args.domain, "domain", DEFAULT_DOMAIN_PATH, none_is_valid=True
     )
-    config = args.config or DEFAULT_CONFIG_PATH
+
+    config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS)
 
     training_files = [
         get_validated_path(f, "data", DEFAULT_DATA_PATH, none_is_valid=True)
@@ -94,7 +108,7 @@ def train_core(
         if isinstance(args.config, list):
             args.config = args.config[0]
 
-        config = args.config or DEFAULT_CONFIG_PATH
+        config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS_CORE)
 
         return train_core(
             domain=args.domain,
@@ -119,7 +133,7 @@ def train_nlu(
 
     output = train_path or args.out
 
-    config = args.config or DEFAULT_CONFIG_PATH
+    config = _get_valid_config(args.config, CONFIG_MANDATORY_KEYS_NLU)
     nlu_data = get_validated_path(
         args.nlu, "nlu", DEFAULT_DATA_PATH, none_is_valid=True
     )
@@ -144,3 +158,48 @@ def extract_additional_arguments(args: argparse.Namespace) -> Dict:
         arguments["debug_plots"] = args.debug_plots
 
     return arguments
+
+
+def _enrich_config(
+    config_path: Text, missing_keys: List[Text], FALLBACK_CONFIG_PATH: Text
+):
+    import rasa.utils.io
+
+    config_data = rasa.utils.io.read_yaml_file(config_path)
+    fallback_config_data = rasa.utils.io.read_yaml_file(FALLBACK_CONFIG_PATH)
+
+    for k in missing_keys:
+        config_data[k] = fallback_config_data[k]
+
+    rasa.utils.io.write_yaml_file(config_data, config_path)
+
+
+def _get_valid_config(
+    config: Optional[Text] = None,
+    default_config: Text = DEFAULT_CONFIG_PATH,
+    mandatory_keys: List[Text] = CONFIG_MANDATORY_KEYS,
+) -> Text:
+    if config:
+        # config is provided via '-c' argument
+        missing_keys = missing_config_keys(config, mandatory_keys)
+        if missing_keys:
+            print_error(
+                "Configuration file '{}' is missing mandatory parameters: "
+                "{}. Add missing parameters to configuration file and try again."
+                "".format(config, ", ".join(missing_keys))
+            )
+            exit(1)
+        return config
+
+    config_path = get_validated_path(default_config, "config", FALLBACK_CONFIG_PATH)
+    missing_keys = missing_config_keys(config_path, mandatory_keys)
+
+    if missing_keys:
+        print_warning(
+            "Configuration file '{}' is missing mandatory parameters: "
+            "{}. Filling missing parameters from fallback configuration file: '{}'."
+            "".format(config, ", ".join(missing_keys), FALLBACK_CONFIG_PATH)
+        )
+        _enrich_config(config_path, missing_keys, FALLBACK_CONFIG_PATH)
+
+    return config_path
