+++ tests/cli/test_rasa_train.py
@@ -1,6 +1,16 @@
 import os
 import shutil
+import tempfile
 
+import pytest
+
+from rasa.cli.train import _get_valid_config
+from rasa.constants import (
+    CONFIG_MANDATORY_KEYS_CORE,
+    CONFIG_MANDATORY_KEYS,
+    CONFIG_MANDATORY_KEYS_NLU,
+    DEFAULT_CONFIG_PATH,
+)
 from rasa.nlu.utils import list_files
 
 
@@ -169,3 +179,41 @@ def test_train_core_help(run):
 
     for i, line in enumerate(lines):
         assert output.outlines[i] == line
+
+
+@pytest.mark.parametrize(
+    "parameters",
+    [
+        {
+            "config_data": {"language": "en", "pipeline": "supervised"},
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_CORE,
+        },
+        {"config_data": {}, "mandatory_keys": CONFIG_MANDATORY_KEYS},
+        {
+            "config_data": {
+                "policy": ["KerasPolicy", "FallbackPolicy"],
+                "imports": "other-folder",
+            },
+            "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU,
+        },
+        {"config_data": None, "mandatory_keys": CONFIG_MANDATORY_KEYS_NLU},
+    ],
+)
+def test_get_valid_config(parameters):
+    import rasa.utils.io
+
+    if parameters["config_data"] is not None:
+        config_path = os.path.join(tempfile.mkdtemp(), "config.yml")
+        rasa.utils.io.write_yaml_file(parameters["config_data"], config_path)
+    else:
+        config_path = DEFAULT_CONFIG_PATH
+
+    config_path = _get_valid_config(None, config_path, parameters["mandatory_keys"])
+    config_data = rasa.utils.io.read_yaml_file(config_path)
+
+    for k in parameters["mandatory_keys"]:
+        assert k in config_data
+
+    if parameters["config_data"] is not None:
+        for k, v in parameters["config_data"].items():
+            assert config_data[k] == v
